project collaborator status permission added

+++ server_api/src/modules/projects/collaborators/collaborator.service.ts
@@ -8,8 +8,6 @@ import {
     GetCollaboratorsOptions,
     UpdateCollaboratorDto,
 } from "./collaborator.dto";
-import { CollaboratorStatus } from "./collaborator.enum";
-import { CollaboratorPermission } from "./collaborator.permission";
 import { CollaboratorRepository, ICollaboratorRepository } from "./collaborator.repository";
 import { CollaboratorStateMachine } from "./collaborator.state-machine";
 
@@ -17,25 +15,25 @@ export class CollaboratorService {
     private repository: ICollaboratorRepository;
     private projectRepository: IProjectRepository;
     private applicantRepo: IApplicantRepository;
-    private permission: CollaboratorPermission;
+    //private permission: CollaboratorPermission;
 
     constructor(repository: ICollaboratorRepository = new CollaboratorRepository(),
         projectRepository?: IProjectRepository, applicantRepo?: IApplicantRepository) {
         this.repository = repository;
         this.projectRepository = projectRepository || new ProjectRepository();
         this.applicantRepo = applicantRepo || new ApplicantRepository();
-        this.permission = new CollaboratorPermission(this.repository);
+        //this.permission = new CollaboratorPermission(this.repository);
     }
 
     async createCollaborator(dto: CreateCollaboratorDto) {
-        const { applicant, project, applicantId, isLeadPI, status } = dto;
+        const { applicant, project, leadPI: applicantId, isLeadPI, status } = dto;
         const projectDoc = await this.projectRepository.findById(project);
         if (!projectDoc) throw new Error("Project not found");
         if (projectDoc.status !== ProjectStatus.pending) {
             throw new Error("Can not add collaborators on non pending projects.");
         }
         if (String(projectDoc.leadPI) !== applicantId) {
-            throw new Error("Only Lead PI can add collaborators.");
+            throw new Error("User not authorized. Lead PI not found.");
         }
         const appDoc = await this.applicantRepo.findOne({ id: applicant });
         if (!appDoc) throw new Error("Applicant not found");
@@ -51,45 +49,45 @@ export class CollaboratorService {
         return collaborators;
     }
 
+    /*
     async updateCollaborator(dto: UpdateCollaboratorDto) {
         const updated = await this.repository.update(dto.id, dto.data);
         return updated;
     }
+    */
 
-    async changeCollaboratorStatus(dto: UpdateCollaboratorDto) {
-        const { id, data, applicantId: userId } = dto;
+    async updateStatus(dto: UpdateCollaboratorDto) {
+        const { id, data, applicantId } = dto;
         const collaboratorDoc = await this.repository.findById(id);
         if (!collaboratorDoc) throw new Error("Collaborator not found");
-        const projectDoc = await this.projectRepository.findById(String(collaboratorDoc.project));
-        if (!projectDoc) throw new Error("Project not found");
         const nextState = data.status;
         if (!nextState) throw new Error("Status is required");
+        if (String(collaboratorDoc.applicant) !== applicantId) {
+            throw new Error(`User not authorized to perform ${nextState}`);
+        }
         const current = collaboratorDoc.status;
-        //if project state is deny here 
-
         // --- State Machine Validation ---
         CollaboratorStateMachine.validateTransition(current, nextState);
-        // Permissions
-        const isActivationChange = current === CollaboratorStatus.verify || nextState === CollaboratorStatus.verify;
-        if (isActivationChange) {
-            await this.permission.validateCollaboratorPermission(id, userId, collaboratorDoc);
-        }
         const updated = await this.repository.update(dto.id, dto.data);
         return updated;
     }
 
-    async deleteCollaborator(dto: DeleteDto) {
+    async delete(dto: DeleteDto) {
         const { id, userId } = dto;
-
-        // Verify user can delete this collaborator
-        //const canDelete = await repository.verifyUserCanDelete(id, userId);
-        //  if (!canDelete) {
-        //throw new Error("User not authorized to delete collaborator or collaborator is active");
-        // }
-
-        const deletedCollaborator = await this.repository.delete(id);
-        if (!deletedCollaborator) throw new Error("Collaborator not found");
-
-        return deletedCollaborator;
+        
+        const collaboratorDoc = await this.repository.findById(id);
+        if (!collaboratorDoc) throw new Error("Collaborator not found");
+        
+        const projectDoc = await this.projectRepository.findById(String(collaboratorDoc.project));
+        if (!projectDoc) throw new Error("Project not found");
+        if (projectDoc.status !== ProjectStatus.pending) {
+            throw new Error("Can not remove collaborators on non pending projects.");
+        }
+        if (String(projectDoc.leadPI) !== userId) {
+            throw new Error("User not authorized. Lead PI not found.");
+        }
+        
+        const deleted = await this.repository.delete(id);
+        return deleted;
     }
 }
